WorkerApp PR feedback: move API client logic to separate class, use updated configs, etc
