+++ airbyte-workers/src/main/java/io/airbyte/workers/WorkerApiClientFactory.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.workers;
+
+import io.airbyte.api.client.AirbyteApiClient;
+
+public interface WorkerApiClientFactory {
+
+  AirbyteApiClient create();
+
+}
